rename Android package/paths
