+++ thermion_flutter/thermion_flutter/android/src/main/kotlin/dev/thermion/android/FilamentInterop.kt
@@ -1,4 +1,4 @@
-package app.polyvox.filament 
+package dev.thermion.android 
 
 import com.sun.jna.ptr.PointerByReference
 import com.sun.jna.ptr.IntByReference
